operator: move files into own Go package
Moves the Go files of the cilium-operator into its own Go package. This
unifies the structure of the operator sources with the other
command-line interfaces in the source, like the cilium-agent.

In addition, this also allows downstream consumers to import the
operator via Go modules (e.g. unit testing via the control-plane
framework).
