+++ operator/cmd/allocator_providers.go
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: Apache-2.0
 // Copyright Authors of Cilium
 
-package main
+package cmd
 
 import (
 	"github.com/cilium/cilium/pkg/ipam/allocator"
