+++ operator/cmd/ccnp_event.go
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: Apache-2.0
 // Copyright Authors of Cilium
 
-package main
+package cmd
 
 import (
 	"context"
@@ -9,88 +9,82 @@ import (
 
 	v1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/fields"
-	"k8s.io/apimachinery/pkg/util/runtime"
 	"k8s.io/apimachinery/pkg/util/wait"
 	"k8s.io/client-go/tools/cache"
 
+	operatorOption "github.com/cilium/cilium/operator/option"
 	"github.com/cilium/cilium/pkg/controller"
 	"github.com/cilium/cilium/pkg/k8s"
 	cilium_v2 "github.com/cilium/cilium/pkg/k8s/apis/cilium.io/v2"
-	v2 "github.com/cilium/cilium/pkg/k8s/apis/cilium.io/v2"
 	"github.com/cilium/cilium/pkg/k8s/informer"
 	"github.com/cilium/cilium/pkg/k8s/watchers/resources"
 	"github.com/cilium/cilium/pkg/kvstore/store"
+	"github.com/cilium/cilium/pkg/metrics"
 	"github.com/cilium/cilium/pkg/option"
 	"github.com/cilium/cilium/pkg/policy/groups"
 )
 
-var (
-	// cnpStatusUpdateInterval is the amount of time between status updates
-	// being sent to the K8s apiserver for a given CNP.
-	cnpStatusUpdateInterval time.Duration
-)
-
-func init() {
-	runtime.ErrorHandlers = []func(error){
-		k8s.K8sErrorHandler,
-	}
+func k8sEventMetric(scope, action string) {
+	metrics.EventTS.WithLabelValues(metrics.LabelEventSourceK8s, scope, action).SetToCurrentTime()
 }
 
-// enableCNPWatcher waits for the CiliumNetowrkPolicy CRD availability and then
-// garbage collects stale CiliumNetowrkPolicy status field entries.
-func enableCNPWatcher() error {
+// enableCCNPWatcher is similar to enableCNPWatcher but handles the watch events for
+// clusterwide policies. Since, internally Clusterwide policies are implemented
+// using CiliumNetworkPolicy itself, the entire implementation uses the methods
+// associcated with CiliumNetworkPolicy.
+func enableCCNPWatcher() error {
 	enableCNPStatusUpdates := kvstoreEnabled() && option.Config.K8sEventHandover && !option.Config.DisableCNPStatusUpdates
 	if enableCNPStatusUpdates {
-		log.Info("Starting CNP Status handover from kvstore to k8s")
+		log.Info("Starting a CCNP Status handover from kvstore to k8s")
 	}
-	log.Info("Starting CNP derivative handler")
+	log.Info("Starting CCNP derivative handler")
 
 	var (
-		cnpStatusMgr *k8s.CNPStatusEventHandler
+		ccnpStatusMgr *k8s.CCNPStatusEventHandler
 	)
-	cnpStore := cache.NewStore(cache.DeletionHandlingMetaNamespaceKeyFunc)
+	ccnpStore := cache.NewStore(cache.DeletionHandlingMetaNamespaceKeyFunc)
 
 	if enableCNPStatusUpdates {
-		cnpStatusMgr = k8s.NewCNPStatusEventHandler(cnpStore, cnpStatusUpdateInterval)
-		cnpSharedStore, err := store.JoinSharedStore(store.Configuration{
-			Prefix: k8s.CNPStatusesPath,
+		ccnpStatusMgr = k8s.NewCCNPStatusEventHandler(ccnpStore, operatorOption.Config.CNPStatusUpdateInterval)
+		ccnpSharedStore, err := store.JoinSharedStore(store.Configuration{
+			Prefix: k8s.CCNPStatusesPath,
 			KeyCreator: func() store.Key {
 				return &k8s.CNPNSWithMeta{}
 			},
-			Observer: cnpStatusMgr,
+			Observer: ccnpStatusMgr,
 		})
 		if err != nil {
 			return err
 		}
 
-		// It is safe to update the CNP store here given the CNP Store
+		// It is safe to update the CCNP store here given the CCNP Store
 		// will only be used by StartStatusHandler method which is used in the
 		// cilium v2 controller below.
-		cnpStatusMgr.UpdateCNPStore(cnpSharedStore)
+		ccnpStatusMgr.UpdateCNPStore(ccnpSharedStore)
 	}
 
 	ciliumV2Controller := informer.NewInformerWithStore(
 		cache.NewListWatchFromClient(k8s.CiliumClient().CiliumV2().RESTClient(),
-			v2.CNPPluralName, v1.NamespaceAll, fields.Everything()),
-		&cilium_v2.CiliumNetworkPolicy{},
+			cilium_v2.CCNPPluralName, v1.NamespaceAll, fields.Everything()),
+		&cilium_v2.CiliumClusterwideNetworkPolicy{},
 		0,
 		cache.ResourceEventHandlerFuncs{
 			AddFunc: func(obj interface{}) {
-				k8sEventMetric(resources.MetricCNP, resources.MetricCreate)
+				k8sEventMetric(resources.MetricCCNP, resources.MetricCreate)
 				if cnp := k8s.ObjToSlimCNP(obj); cnp != nil {
 					// We need to deepcopy this structure because we are writing
 					// fields.
 					// See https://github.com/cilium/cilium/blob/27fee207f5422c95479422162e9ea0d2f2b6c770/pkg/policy/api/ingress.go#L112-L134
 					cnpCpy := cnp.DeepCopy()
 
-					groups.AddDerivativeCNPIfNeeded(cnpCpy.CiliumNetworkPolicy)
+					groups.AddDerivativeCCNPIfNeeded(cnpCpy.CiliumNetworkPolicy)
 					if enableCNPStatusUpdates {
-						cnpStatusMgr.StartStatusHandler(cnpCpy)
+						ccnpStatusMgr.StartStatusHandler(cnpCpy)
 					}
 				}
 			},
 			UpdateFunc: func(oldObj, newObj interface{}) {
-				k8sEventMetric(resources.MetricCNP, resources.MetricUpdate)
+				k8sEventMetric(resources.MetricCCNP, resources.MetricUpdate)
 				if oldCNP := k8s.ObjToSlimCNP(oldObj); oldCNP != nil {
 					if newCNP := k8s.ObjToSlimCNP(newObj); newCNP != nil {
 						if oldCNP.DeepEqual(newCNP) {
@@ -103,12 +97,12 @@ func enableCNPWatcher() error {
 						newCNPCpy := newCNP.DeepCopy()
 						oldCNPCpy := oldCNP.DeepCopy()
 
-						groups.UpdateDerivativeCNPIfNeeded(newCNPCpy.CiliumNetworkPolicy, oldCNPCpy.CiliumNetworkPolicy)
+						groups.UpdateDerivativeCCNPIfNeeded(newCNPCpy.CiliumNetworkPolicy, oldCNPCpy.CiliumNetworkPolicy)
 					}
 				}
 			},
 			DeleteFunc: func(obj interface{}) {
-				k8sEventMetric(resources.MetricCNP, resources.MetricDelete)
+				k8sEventMetric(resources.MetricCCNP, resources.MetricDelete)
 				cnp := k8s.ObjToSlimCNP(obj)
 				if cnp == nil {
 					return
@@ -117,16 +111,16 @@ func enableCNPWatcher() error {
 				// to delete the cnp from the pooling.
 				groups.DeleteDerivativeFromCache(cnp.CiliumNetworkPolicy)
 				if enableCNPStatusUpdates {
-					cnpStatusMgr.StopStatusHandler(cnp)
+					ccnpStatusMgr.StopStatusHandler(cnp)
 				}
 			},
 		},
-		k8s.ConvertToCNP,
-		cnpStore,
+		k8s.ConvertToCCNP,
+		ccnpStore,
 	)
 	go ciliumV2Controller.Run(wait.NeverStop)
 
-	controller.NewManager().UpdateController("cnp-to-groups",
+	controller.NewManager().UpdateController("ccnp-to-groups",
 		controller.ControllerParams{
 			DoFunc: func(ctx context.Context) error {
 				groups.UpdateCNPInformation()
