+++ operator/cmd/cnp_event.go
@@ -1,90 +1,96 @@
 // SPDX-License-Identifier: Apache-2.0
 // Copyright Authors of Cilium
 
-package main
+package cmd
 
 import (
 	"context"
 	"time"
 
+	v1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/fields"
+	"k8s.io/apimachinery/pkg/util/runtime"
 	"k8s.io/apimachinery/pkg/util/wait"
 	"k8s.io/client-go/tools/cache"
 
-	operatorOption "github.com/cilium/cilium/operator/option"
 	"github.com/cilium/cilium/pkg/controller"
 	"github.com/cilium/cilium/pkg/k8s"
 	cilium_v2 "github.com/cilium/cilium/pkg/k8s/apis/cilium.io/v2"
+	v2 "github.com/cilium/cilium/pkg/k8s/apis/cilium.io/v2"
 	"github.com/cilium/cilium/pkg/k8s/informer"
-	v1 "github.com/cilium/cilium/pkg/k8s/slim/k8s/apis/meta/v1"
 	"github.com/cilium/cilium/pkg/k8s/watchers/resources"
 	"github.com/cilium/cilium/pkg/kvstore/store"
-	"github.com/cilium/cilium/pkg/metrics"
 	"github.com/cilium/cilium/pkg/option"
 	"github.com/cilium/cilium/pkg/policy/groups"
 )
 
-func k8sEventMetric(scope, action string) {
-	metrics.EventTS.WithLabelValues(metrics.LabelEventSourceK8s, scope, action).SetToCurrentTime()
+var (
+	// cnpStatusUpdateInterval is the amount of time between status updates
+	// being sent to the K8s apiserver for a given CNP.
+	cnpStatusUpdateInterval time.Duration
+)
+
+func init() {
+	runtime.ErrorHandlers = []func(error){
+		k8s.K8sErrorHandler,
+	}
 }
 
-// enableCCNPWatcher is similar to enableCNPWatcher but handles the watch events for
-// clusterwide policies. Since, internally Clusterwide policies are implemented
-// using CiliumNetworkPolicy itself, the entire implementation uses the methods
-// associcated with CiliumNetworkPolicy.
-func enableCCNPWatcher() error {
+// enableCNPWatcher waits for the CiliumNetowrkPolicy CRD availability and then
+// garbage collects stale CiliumNetowrkPolicy status field entries.
+func enableCNPWatcher() error {
 	enableCNPStatusUpdates := kvstoreEnabled() && option.Config.K8sEventHandover && !option.Config.DisableCNPStatusUpdates
 	if enableCNPStatusUpdates {
-		log.Info("Starting a CCNP Status handover from kvstore to k8s")
+		log.Info("Starting CNP Status handover from kvstore to k8s")
 	}
-	log.Info("Starting CCNP derivative handler")
+	log.Info("Starting CNP derivative handler")
 
 	var (
-		ccnpStatusMgr *k8s.CCNPStatusEventHandler
+		cnpStatusMgr *k8s.CNPStatusEventHandler
 	)
-	ccnpStore := cache.NewStore(cache.DeletionHandlingMetaNamespaceKeyFunc)
+	cnpStore := cache.NewStore(cache.DeletionHandlingMetaNamespaceKeyFunc)
 
 	if enableCNPStatusUpdates {
-		ccnpStatusMgr = k8s.NewCCNPStatusEventHandler(ccnpStore, operatorOption.Config.CNPStatusUpdateInterval)
-		ccnpSharedStore, err := store.JoinSharedStore(store.Configuration{
-			Prefix: k8s.CCNPStatusesPath,
+		cnpStatusMgr = k8s.NewCNPStatusEventHandler(cnpStore, cnpStatusUpdateInterval)
+		cnpSharedStore, err := store.JoinSharedStore(store.Configuration{
+			Prefix: k8s.CNPStatusesPath,
 			KeyCreator: func() store.Key {
 				return &k8s.CNPNSWithMeta{}
 			},
-			Observer: ccnpStatusMgr,
+			Observer: cnpStatusMgr,
 		})
 		if err != nil {
 			return err
 		}
 
-		// It is safe to update the CCNP store here given the CCNP Store
+		// It is safe to update the CNP store here given the CNP Store
 		// will only be used by StartStatusHandler method which is used in the
 		// cilium v2 controller below.
-		ccnpStatusMgr.UpdateCNPStore(ccnpSharedStore)
+		cnpStatusMgr.UpdateCNPStore(cnpSharedStore)
 	}
 
 	ciliumV2Controller := informer.NewInformerWithStore(
 		cache.NewListWatchFromClient(k8s.CiliumClient().CiliumV2().RESTClient(),
-			cilium_v2.CCNPPluralName, v1.NamespaceAll, fields.Everything()),
-		&cilium_v2.CiliumClusterwideNetworkPolicy{},
+			v2.CNPPluralName, v1.NamespaceAll, fields.Everything()),
+		&cilium_v2.CiliumNetworkPolicy{},
 		0,
 		cache.ResourceEventHandlerFuncs{
 			AddFunc: func(obj interface{}) {
-				k8sEventMetric(resources.MetricCCNP, resources.MetricCreate)
+				k8sEventMetric(resources.MetricCNP, resources.MetricCreate)
 				if cnp := k8s.ObjToSlimCNP(obj); cnp != nil {
 					// We need to deepcopy this structure because we are writing
 					// fields.
 					// See https://github.com/cilium/cilium/blob/27fee207f5422c95479422162e9ea0d2f2b6c770/pkg/policy/api/ingress.go#L112-L134
 					cnpCpy := cnp.DeepCopy()
 
-					groups.AddDerivativeCCNPIfNeeded(cnpCpy.CiliumNetworkPolicy)
+					groups.AddDerivativeCNPIfNeeded(cnpCpy.CiliumNetworkPolicy)
 					if enableCNPStatusUpdates {
-						ccnpStatusMgr.StartStatusHandler(cnpCpy)
+						cnpStatusMgr.StartStatusHandler(cnpCpy)
 					}
 				}
 			},
 			UpdateFunc: func(oldObj, newObj interface{}) {
-				k8sEventMetric(resources.MetricCCNP, resources.MetricUpdate)
+				k8sEventMetric(resources.MetricCNP, resources.MetricUpdate)
 				if oldCNP := k8s.ObjToSlimCNP(oldObj); oldCNP != nil {
 					if newCNP := k8s.ObjToSlimCNP(newObj); newCNP != nil {
 						if oldCNP.DeepEqual(newCNP) {
@@ -97,12 +103,12 @@ func enableCCNPWatcher() error {
 						newCNPCpy := newCNP.DeepCopy()
 						oldCNPCpy := oldCNP.DeepCopy()
 
-						groups.UpdateDerivativeCCNPIfNeeded(newCNPCpy.CiliumNetworkPolicy, oldCNPCpy.CiliumNetworkPolicy)
+						groups.UpdateDerivativeCNPIfNeeded(newCNPCpy.CiliumNetworkPolicy, oldCNPCpy.CiliumNetworkPolicy)
 					}
 				}
 			},
 			DeleteFunc: func(obj interface{}) {
-				k8sEventMetric(resources.MetricCCNP, resources.MetricDelete)
+				k8sEventMetric(resources.MetricCNP, resources.MetricDelete)
 				cnp := k8s.ObjToSlimCNP(obj)
 				if cnp == nil {
 					return
@@ -111,16 +117,16 @@ func enableCCNPWatcher() error {
 				// to delete the cnp from the pooling.
 				groups.DeleteDerivativeFromCache(cnp.CiliumNetworkPolicy)
 				if enableCNPStatusUpdates {
-					ccnpStatusMgr.StopStatusHandler(cnp)
+					cnpStatusMgr.StopStatusHandler(cnp)
 				}
 			},
 		},
-		k8s.ConvertToCCNP,
-		ccnpStore,
+		k8s.ConvertToCNP,
+		cnpStore,
 	)
 	go ciliumV2Controller.Run(wait.NeverStop)
 
-	controller.NewManager().UpdateController("ccnp-to-groups",
+	controller.NewManager().UpdateController("cnp-to-groups",
 		controller.ControllerParams{
 			DoFunc: func(ctx context.Context) error {
 				groups.UpdateCNPInformation()
