+++ operator/cmd/provider_alibabacloud_flags.go
@@ -3,7 +3,7 @@
 
 //go:build ipam_provider_alibabacloud
 
-package main
+package cmd
 
 import (
 	"github.com/spf13/viper"
