+++ operator/cmd/provider_alibabacloud_register.go
@@ -3,7 +3,7 @@
 
 //go:build ipam_provider_alibabacloud
 
-package main
+package cmd
 
 import (
 	"github.com/cilium/cilium/pkg/ipam/allocator/alibabacloud"
