+++ operator/cmd/provider_azure_flags.go
@@ -3,7 +3,7 @@
 
 //go:build ipam_provider_azure
 
-package main
+package cmd
 
 import (
 	"github.com/spf13/viper"
