+++ operator/cmd/provider_operator_register.go
@@ -3,7 +3,7 @@
 
 //go:build ipam_provider_operator
 
-package main
+package cmd
 
 import (
 	// These dependencies should be included only when this file is included in the build.
